requirements.txt
Django>=5.0
psycopg[binary]>=3.2
djangorestframework>=3.15
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.3
django-filter>=24.2
redis>=5.0
celery>=5.4
python-dotenv>=1.0
whitenoise>=6.7
django-cors-headers>=4.3